Update the implementation of ttl::evaluate for builtin scalars.
I want the evaluate function to act like an identity function with respect to
types and values, except that if the input is an rvalue reference I want to
return the scalar by value.

Essentially I want the following deductions.

```
T         | evaluate(T&&)
----------+--------------
T&        | T&
T const&  | T const&
T&&       | T
T const&& | T const
```

The correct way to do this for a forwarding reference is `-> T`. This is because
of how forwarding reference deduction happens. For T&&, T is deduced as an
lvalue reference for lvalues, and as the valut type for rvalues (in which case
the parameter type is the rvalue reference T&&).