Don't make expressions non-copy-movable.
This might be something I want to try and do someday, but it's going to cause a
lot of complications and present weird lifetime issues during development that I
don't want to have to worry about. In particular, it will trigger reference
lifetime extension and that's going to be annoying to debug.

Honestly I'm not even sure it's possible and I don't want to experimentally both
the DSL and this form of expression at the same time.