Assert that contracted indices in products have compatible extents.
Implicit summation means that a contracted index is one that is involved in a
dot product, and such dot products must be over extents that have the same
dimension.

This adds a static check to make sure that contracted extents are at least
compatible (i.e., if they are both static extents, then the size of each extent
is the same).